LayoutTests/platform/efl/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left is 8
PASS rects[0].top is 8
PASS rects[0].width is 400
PASS rects[0].height is 160
PASS rects[1].left is 8
PASS rects[1].top is 19
FAIL rects[1].width should be 387. Was 396.
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 59
FAIL rects[2].width should be 360. Was 398.
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
PASS rects[3].top is 99
FAIL rects[3].width should be 382. Was 360.
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 139
FAIL rects[4].width should be 369. Was 306.
FAIL rects[4].height should be 18. Was 17.
Test 2
PASS rects.length is 4
PASS rects[0].left is 8
PASS rects[0].top is 215
FAIL rects[0].width should be 387. Was 396.
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
PASS rects[1].top is 255
FAIL rects[1].width should be 360. Was 398.
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 295
FAIL rects[2].width should be 382. Was 360.
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
PASS rects[3].top is 335
FAIL rects[3].width should be 369. Was 306.
FAIL rects[3].height should be 18. Was 17.
Test 3
PASS rects.length is 4
FAIL rects[0].left should be 81. Was 80.
PASS rects[0].top is 411
FAIL rects[0].width should be 314. Was 324.
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
PASS rects[1].top is 451
FAIL rects[1].width should be 360. Was 398.
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 491
FAIL rects[2].width should be 382. Was 360.
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
PASS rects[3].top is 531
FAIL rects[3].width should be 147. Was 88.
FAIL rects[3].height should be 18. Was 17.
Test 4
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 596
PASS rects[0].width is 400
PASS rects[0].height is 278
PASS rects[1].left is 8
PASS rects[1].top is 607
FAIL rects[1].width should be 387. Was 396.
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 725
FAIL rects[2].width should be 269. Was 242.
FAIL rects[2].height should be 18. Was 17.
FAIL rects[3].left should be 352. Was 326.
PASS rects[3].top is 725
FAIL rects[3].width should be 46. Was 44.
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 765
FAIL rects[4].width should be 349. Was 341.
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
PASS rects[5].top is 805
FAIL rects[5].width should be 376. Was 366.
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
PASS rects[6].top is 845
FAIL rects[6].width should be 68. Was 67.
FAIL rects[6].height should be 18. Was 17.
Test 5
PASS rects.length is 5
FAIL rects[0].left should be 81. Was 80.
PASS rects[0].top is 921
FAIL rects[0].width should be 314. Was 324.
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
PASS rects[1].top is 1039
FAIL rects[1].width should be 269. Was 242.
FAIL rects[1].height should be 18. Was 17.
FAIL rects[2].left should be 276. Was 250.
PASS rects[2].top is 950
FAIL rects[2].width should be 77. Was 76.
PASS rects[2].height is 103
FAIL rects[3].left should be 352. Was 326.
PASS rects[3].top is 1039
FAIL rects[3].width should be 46. Was 44.
FAIL rects[3].height should be 18. Was 17.
PASS rects[4].left is 8
PASS rects[4].top is 1079
FAIL rects[4].width should be 291. Was 284.
FAIL rects[4].height should be 18. Was 17.
Test 6
FAIL rects.length should be 8. Was 7.
PASS rects[0].left is 8
PASS rects[0].top is 1224
PASS rects[0].width is 400
FAIL rects[0].height should be 240. Was 200.
PASS rects[1].left is 8
PASS rects[1].top is 1235
FAIL rects[1].width should be 387. Was 396.
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
PASS rects[2].top is 1275
FAIL rects[2].width should be 16. Was 57.
FAIL rects[2].height should be 18. Was 17.
FAIL rects[3].left should be 8. Was 105.
FAIL rects[3].top should be 1315. Was 1275.
FAIL rects[3].width should be 59. Was 44.
FAIL rects[3].height should be 18. Was 17.
FAIL rects[4].left should be 106. Was 8.
PASS rects[4].top is 1315
FAIL rects[4].width should be 46. Was 387.
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
PASS rects[5].top is 1355
PASS rects[5].width is 397
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
PASS rects[6].top is 1395
FAIL rects[6].width should be 360. Was 171.
FAIL rects[6].height should be 18. Was 17.
FAIL rects[7].left should be 8. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].left')
FAIL rects[7].top should be 1435. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].top')
FAIL rects[7].width should be 222. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].width')
FAIL rects[7].height should be 18. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].height')
Test 7
FAIL rects.length should be 8. Was 7.
PASS rects[0].left is 8
FAIL rects[0].top should be 1511. Was 1471.
FAIL rects[0].width should be 387. Was 396.
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
FAIL rects[1].top should be 1551. Was 1500.
FAIL rects[1].width should be 16. Was 400.
FAIL rects[1].height should be 18. Was 40.
PASS rects[2].left is 8
FAIL rects[2].top should be 1580. Was 1511.
FAIL rects[2].width should be 400. Was 57.
FAIL rects[2].height should be 40. Was 17.
FAIL rects[3].left should be 8. Was 105.
FAIL rects[3].top should be 1591. Was 1511.
FAIL rects[3].width should be 59. Was 44.
FAIL rects[3].height should be 18. Was 17.
FAIL rects[4].left should be 106. Was 8.
FAIL rects[4].top should be 1591. Was 1551.
FAIL rects[4].width should be 46. Was 387.
FAIL rects[4].height should be 18. Was 17.
PASS rects[5].left is 8
FAIL rects[5].top should be 1631. Was 1591.
PASS rects[5].width is 397
FAIL rects[5].height should be 18. Was 17.
PASS rects[6].left is 8
FAIL rects[6].top should be 1671. Was 1631.
FAIL rects[6].width should be 360. Was 171.
FAIL rects[6].height should be 18. Was 17.
FAIL rects[7].left should be 8. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].left')
FAIL rects[7].top should be 1711. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].top')
FAIL rects[7].width should be 222. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].width')
FAIL rects[7].height should be 18. Threw exception TypeError: undefined is not an object (evaluating 'rects[7].height')
Test 8
PASS rects.length is 4
FAIL rects[0].left should be 168. Was 169.
FAIL rects[0].top should be 1758. Was 1678.
FAIL rects[0].width should be 263. Was 268.
FAIL rects[0].height should be 309. Was 315.
FAIL rects[1].left should be 137. Was 138.
FAIL rects[1].top should be 1784. Was 1704.
FAIL rects[1].width should be 247. Was 270.
FAIL rects[1].height should be 288. Was 316.
PASS rects[2].left is 107
FAIL rects[2].top should be 1809. Was 1729.
FAIL rects[2].width should be 260. Was 246.
FAIL rects[2].height should be 305. Was 288.
FAIL rects[3].left should be 76. Was 77.
FAIL rects[3].top should be 1835. Was 1755.
FAIL rects[3].width should be 252. Was 211.
FAIL rects[3].height should be 295. Was 246.
Test 9
PASS rects.length is 0
Test 9b
PASS rects.length is 1
PASS rects[0].left is 8
FAIL rects[0].top should be 1983. Was 1903.
PASS rects[0].width is 0
FAIL rects[0].height should be 18. Was 17.
Test 10
FAIL rects.length should be 5. Was 4.
PASS rects[0].left is 19
FAIL rects[0].top should be 2168. Was 2088.
FAIL rects[0].width should be 18. Was 17.
FAIL rects[0].height should be 375. Was 393.
PASS rects[1].left is 59
FAIL rects[1].top should be 2168. Was 2088.
FAIL rects[1].width should be 18. Was 17.
FAIL rects[1].height should be 368. Was 377.
PASS rects[2].left is 99
FAIL rects[2].top should be 2168. Was 2088.
FAIL rects[2].width should be 18. Was 17.
FAIL rects[2].height should be 349. Was 372.
PASS rects[3].left is 139
FAIL rects[3].top should be 2168. Was 2088.
FAIL rects[3].width should be 18. Was 17.
FAIL rects[3].height should be 376. Was 360.
FAIL rects[4].left should be 179. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].left')
FAIL rects[4].top should be 2168. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].top')
FAIL rects[4].width should be 18. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].width')
FAIL rects[4].height should be 68. Threw exception TypeError: undefined is not an object (evaluating 'rects[4].height')
Test 11
PASS rects.length is 2
PASS rects[0].left is 8
FAIL rects[0].top should be 2604. Was 2524.
PASS rects[0].width is 400
PASS rects[0].height is 40
PASS rects[1].left is 8
FAIL rects[1].top should be 2615. Was 2535.
FAIL rects[1].width should be 182. Was 177.
FAIL rects[1].height should be 18. Was 17.
Test 12
PASS rects.length is 1
PASS rects[0].left is 8
FAIL rects[0].top should be 2840. Was 2760.
PASS rects[0].width is 400
PASS rects[0].height is 160
Test 13
PASS rects.length is 4
PASS rects[0].left is 8
FAIL rects[0].top should be 3047. Was 2967.
FAIL rects[0].width should be 387. Was 396.
FAIL rects[0].height should be 18. Was 17.
PASS rects[1].left is 8
FAIL rects[1].top should be 3087. Was 3007.
FAIL rects[1].width should be 360. Was 398.
FAIL rects[1].height should be 18. Was 17.
PASS rects[2].left is 8
FAIL rects[2].top should be 3127. Was 3047.
FAIL rects[2].width should be 382. Was 360.
FAIL rects[2].height should be 18. Was 17.
PASS rects[3].left is 8
FAIL rects[3].top should be 3167. Was 3087.
FAIL rects[3].width should be 369. Was 306.
FAIL rects[3].height should be 18. Was 17.
Test 14
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
